Fix title cut off in week view for ~50 minute event
If we are in week view and have an event around ~50 minutes long, it
didn't apply the 'fc-short' tag to the element, but it also did not have
the 'bottom' property set big enough to fully display the title
underneath the time. This commit fixes that, as well as displaying the
title of an event (instead of only the time of an event) for a 45 minute
time block, which is the more important piece of information.